+++ sdk/core/azure-core-http-netty/src/test/java/com/azure/core/http/netty/ReactorNettyClientTests.java
@@ -7,10 +7,15 @@ import com.azure.core.http.HttpClient;
 import com.azure.core.http.HttpMethod;
 import com.azure.core.http.HttpRequest;
 import com.azure.core.http.HttpResponse;
+import com.azure.core.http.netty.implementation.ReactorNettyClientProvider;
+import com.azure.core.util.Context;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
 import io.netty.buffer.ByteBuf;
+import java.io.ByteArrayOutputStream;
+import java.util.ArrayList;
+import java.util.List;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
@@ -37,9 +42,14 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import static com.azure.core.http.netty.NettyAsyncHttpClient.ReactorNettyHttpResponse;
 import static java.time.Duration.ofMillis;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTimeout;
 
 public class ReactorNettyClientTests {
+    private static final String SHORT_BODY_PATH = "/short";
+    private static final String LONG_BODY_PATH = "/long";
 
     private static final String SHORT_BODY = "hi there";
     private static final String LONG_BODY = createLongBody();
@@ -50,8 +60,8 @@ public class ReactorNettyClientTests {
     public static void beforeClass() {
         server = new WireMockServer(WireMockConfiguration.options().dynamicPort().disableRequestJournal());
         server.stubFor(
-                WireMock.get("/short").willReturn(WireMock.aResponse().withBody(SHORT_BODY)));
-        server.stubFor(WireMock.get("/long").willReturn(WireMock.aResponse().withBody(LONG_BODY)));
+                WireMock.get(SHORT_BODY_PATH).willReturn(WireMock.aResponse().withBody(SHORT_BODY)));
+        server.stubFor(WireMock.get(LONG_BODY_PATH).willReturn(WireMock.aResponse().withBody(LONG_BODY)));
         server.stubFor(WireMock.get("/error")
                 .willReturn(WireMock.aResponse().withBody("error").withStatus(500)));
         server.stubFor(
@@ -69,17 +79,17 @@ public class ReactorNettyClientTests {
 
     @Test
     public void testFlowableResponseShortBodyAsByteArrayAsync() {
-        checkBodyReceived(SHORT_BODY, "/short");
+        checkBodyReceived(SHORT_BODY, SHORT_BODY_PATH);
     }
 
     @Test
     public void testFlowableResponseLongBodyAsByteArrayAsync() {
-        checkBodyReceived(LONG_BODY, "/long");
+        checkBodyReceived(LONG_BODY, LONG_BODY_PATH);
     }
 
     @Test
     public void testMultipleSubscriptionsEmitsError() {
-        HttpResponse response = getResponse("/short");
+        HttpResponse response = getResponse(SHORT_BODY_PATH);
         // Subscription:1
         response.getBodyAsByteArray().block();
 
@@ -94,7 +104,7 @@ public class ReactorNettyClientTests {
 
     @Test
     public void testDispose() throws InterruptedException {
-        ReactorNettyHttpResponse response = getResponse("/long");
+        ReactorNettyHttpResponse response = getResponse(LONG_BODY_PATH);
         response.getBody().subscribe().dispose();
         // Wait for scheduled connection disposal action to execute on netty event-loop
         Thread.sleep(5000);
@@ -103,7 +113,7 @@ public class ReactorNettyClientTests {
 
     @Test
     public void testCancel() {
-        ReactorNettyHttpResponse response = getResponse("/long");
+        ReactorNettyHttpResponse response = getResponse(LONG_BODY_PATH);
         //
         StepVerifierOptions stepVerifierOptions = StepVerifierOptions.create();
         stepVerifierOptions.initialRequest(0);
@@ -129,7 +139,7 @@ public class ReactorNettyClientTests {
     @Test
     @Disabled("Not working accurately at present")
     public void testFlowableBackpressure() {
-        HttpResponse response = getResponse("/long");
+        HttpResponse response = getResponse(LONG_BODY_PATH);
         //
         StepVerifierOptions stepVerifierOptions = StepVerifierOptions.create();
         stepVerifierOptions.initialRequest(0);
@@ -235,7 +245,7 @@ public class ReactorNettyClientTests {
 //        Mono<Long> numBytesMono = Flux.range(1, numRequests)
 //                .parallel(10)
 //                .runOn(reactor.core.scheduler.Schedulers.newElastic("io", 30))
-//                .flatMap(n -> Mono.fromCallable(() -> getResponse(client, "/long")).flatMapMany(response -> {
+//                .flatMap(n -> Mono.fromCallable(() -> getResponse(client, LONG_BODY_PATH)).flatMapMany(response -> {
 //                    MessageDigest md = md5Digest();
 //                    return response.body()
 //                            .doOnNext(bb -> {
@@ -280,6 +290,56 @@ public class ReactorNettyClientTests {
 ////        assertEquals(numRequests * LONG_BODY.getBytes(StandardCharsets.UTF_8).length, numBytes);
 
         Assertions.fail("Method needs to be reimplemented");
+    }
+
+    /**
+     * Tests that deep copying the buffers returned by Netty will make the stream returned to the customer resilient to
+     * Netty reclaiming them once the 'onNext' operator chain has completed.
+     */
+    @Test
+    public void deepCopyBufferConfiguredInBuilder() {
+        HttpClient client = new NettyAsyncHttpClientBuilder().disableBufferCopy(false).build();
+
+        HttpResponse response = client.send(new HttpRequest(HttpMethod.GET, url(server, LONG_BODY_PATH))).block();
+        assertNotNull(response);
+        assertEquals(200, response.getStatusCode());
+
+        DelayWriteStream delayWriteStream = new DelayWriteStream();
+        response.getBody().doOnNext(delayWriteStream::write).blockLast();
+        assertEquals(LONG_BODY, delayWriteStream.aggregateAsString());
+    }
+
+    /**
+     * Tests that preventing deep copying the buffers returned by Netty won't make the stream returned to the customer
+     * resilient to Netty reclaiming them once the 'onNext' operator chain has completed.
+     */
+    @Test
+    public void ignoreDeepCopyBufferConfiguredInBuilder() {
+        HttpClient client = new NettyAsyncHttpClientBuilder().disableBufferCopy(true).build();
+
+        HttpResponse response = client.send(new HttpRequest(HttpMethod.GET, url(server, LONG_BODY_PATH))).block();
+        assertNotNull(response);
+        assertEquals(200, response.getStatusCode());
+
+        DelayWriteStream delayWriteStream = new DelayWriteStream();
+        response.getBody().doOnNext(delayWriteStream::write).blockLast();
+        assertNotEquals(LONG_BODY, delayWriteStream.aggregateAsString());
+    }
+
+    /**
+     * Tests that deep copying of buffers is able to be configured via {@link Context}.
+     */
+    @Test
+    public void deepCopyBufferConfiguredByContext() {
+        HttpClient client = new ReactorNettyClientProvider().createInstance();
+
+        HttpResponse response = client.send(new HttpRequest(HttpMethod.GET, url(server, LONG_BODY_PATH))).block();
+        assertNotNull(response);
+        assertEquals(200, response.getStatusCode());
+
+        DelayWriteStream delayWriteStream = new DelayWriteStream();
+        response.getBody().doOnNext(delayWriteStream::write).blockLast();
+        assertEquals(LONG_BODY, delayWriteStream.aggregateAsString());
     }
 
     private static MessageDigest md5Digest() {
@@ -345,5 +405,28 @@ public class ReactorNettyClientTests {
         HttpRequest request = new HttpRequest(HttpMethod.GET, url(server, path));
         ReactorNettyHttpResponse response = (ReactorNettyHttpResponse) client.send(request).block();
         return response;
+    }
+
+    private static final class DelayWriteStream {
+        List<ByteBuffer> internalBuffers = new ArrayList<>();
+
+        public void write(ByteBuffer buffer) {
+            internalBuffers.add(buffer);
+        }
+
+        public String aggregateAsString() {
+            ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+
+            for (ByteBuffer buffer : internalBuffers) {
+                int bufferSize = buffer.remaining();
+                int offset = buffer.position();
+
+                for (int i = 0; i < bufferSize; i++) {
+                    outputStream.write(buffer.get(i + offset));
+                }
+            }
+
+            return new String(outputStream.toByteArray(), StandardCharsets.UTF_8);
+        }
     }
 }