+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Release History
+## 12.3.1 (2020-02-10)
+- Updated `azure-core-http-netty` to version 1.3.0
+- Update `azure-storage-common` to version 12.3.1
 
 ## 12.3.0 (2020-01-16)
 This package's