+++ sdk/storage/azure-storage-file-datalake/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 12.0.0-beta.11 (2020-02-10)
+- Updated `azure-core-http-netty` to version 1.3.0
+- Update `azure-storage-blob` to version 12.3.1
+
 ## 12.0.0-beta.10 (2020-01-15)
 This package's
 [documentation](https://github.com/Azure/azure-sdk-for-java/blob/azure-storage-file-datalake_12.0.0-beta.10/sdk/storage/azure-storage-file-datalake/README.md)