+++ sdk/storage/azure-storage-file-share/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 12.1.1 (2020-02-10)
+- Updated `azure-core-http-netty` to version 1.3.0
+- Update `azure-storage-common` to version 12.3.1
+
 ## 12.1.0 (2020-01-08)
 This package's
 [documentation](https://github.com/Azure/azure-sdk-for-java/blob/azure-storage-file_12.1.0/sdk/storage/azure-storage-file-share/README.md)